Compartmentalizing into individual componets
