+++ harbor/app/(tabs)/index.tsx
@@ -1,196 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import { StyleSheet, View } from 'react-native';
-import MapView, { Marker } from 'react-native-maps';
-import Geolocation from 'react-native-geolocation-service';
-import { PROVIDER_GOOGLE } from 'react-native-maps'
-
-// Mapview set up referencing the below documentation
-// https://docs.expo.dev/versions/latest/sdk/map-view/
-
-// Dark mode style JSON
-const darkModeStyle = [
-  {
-    elementType: 'geometry',
-    stylers: [{ color: '#212121' }],
-  },
-  {
-    elementType: 'labels.icon',
-    stylers: [{ visibility: 'off' }],
-  },
-  {
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#757575' }],
-  },
-  {
-    elementType: 'labels.text.stroke',
-    stylers: [{ color: '#212121' }],
-  },
-  {
-    featureType: 'administrative',
-    elementType: 'geometry',
-    stylers: [{ color: '#757575' }],
-  },
-  {
-    featureType: 'administrative.country',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#9e9e9e' }],
-  },
-  {
-    featureType: 'administrative.land_parcel',
-    stylers: [{ visibility: 'off' }],
-  },
-  {
-    featureType: 'administrative.locality',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#bdbdbd' }],
-  },
-  {
-    featureType: 'poi',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#757575' }],
-  },
-  {
-    featureType: 'poi.park',
-    elementType: 'geometry',
-    stylers: [{ color: '#181818' }],
-  },
-  {
-    featureType: 'poi.park',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#616161' }],
-  },
-  {
-    featureType: 'poi.park',
-    elementType: 'labels.text.stroke',
-    stylers: [{ color: '#1b1b1b' }],
-  },
-  {
-    featureType: 'road',
-    elementType: 'geometry.fill',
-    stylers: [{ color: '#2c2c2c' }],
-  },
-  {
-    featureType: 'road',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#8a8a8a' }],
-  },
-  {
-    featureType: 'road.arterial',
-    elementType: 'geometry',
-    stylers: [{ color: '#373737' }],
-  },
-  {
-    featureType: 'road.highway',
-    elementType: 'geometry',
-    stylers: [{ color: '#3c3c3c' }],
-  },
-  {
-    featureType: 'road.highway.controlled_access',
-    elementType: 'geometry',
-    stylers: [{ color: '#4e4e4e' }],
-  },
-  {
-    featureType: 'road.local',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#616161' }],
-  },
-  {
-    featureType: 'transit',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#757575' }],
-  },
-  {
-    featureType: 'water',
-    elementType: 'geometry',
-    stylers: [{ color: '#000000' }],
-  },
-  {
-    featureType: 'water',
-    elementType: 'labels.text.fill',
-    stylers: [{ color: '#3d3d3d' }],
-  },
-];
+import React from 'react';
+import MapViewComponent from '@/components/mapView';
+import { View } from 'react-native';
+import SosPointer from '@/assets/images/pointers/sosPointer';
+import { Marker } from '@/types';
 
 export default function HomeScreen() {
-  const [currentLocation, setCurrentLocation] = useState<{
-    latitude: number;
-    longitude: number;
-    latitudeDelta: number;
-    longitudeDelta: number;
-  } | null>(null);
-  const [markers, setMarkers] = useState([
-    { id: 1, latitude: 37.78825, longitude: -122.4324, title: 'Marker 1' },
-    { id: 2, latitude: 37.78925, longitude: -122.4334, title: 'Marker 2' },
-    { id: 3, latitude: 37.79025, longitude: -122.4344, title: 'Marker 3' },
-  ]);
-
-  const defaultRegion = {
-    latitude: 0,
-    longitude: 0,
-    latitudeDelta: 100,
-    longitudeDelta: 100,
-  };
-
-  useEffect(() => {
-    Geolocation.requestAuthorization('whenInUse').then(() => {
-      Geolocation.getCurrentPosition(
-        (position) => {
-          setCurrentLocation({
-            latitude: position.coords.latitude,
-            longitude: position.coords.longitude,
-            latitudeDelta: 0.01,
-            longitudeDelta: 0.01,
-          });
-        },
-        (error) => {
-          console.error(error);
-          setCurrentLocation(null);
-        },
-        { enableHighAccuracy: true, timeout: 15000, maximumAge: 10000 }
-      );
-    });
-  }, []);
+  const markers: Marker[] = [
+    { id: 1, latitude: 37.78825, longitude: -122.4324, isSOS: true, isVerified: false, date: new Date('2025-04-19T10:00:00Z') },
+    { id: 2, latitude: 37.78925, longitude: -122.4334, isSOS: false, isVerified: false, date: new Date('2025-04-19T10:00:00Z') },
+    { id: 3, latitude: 37.79025, longitude: -122.4344, isSOS: true, isVerified: false, date: new Date('2025-04-19T10:00:00Z') },
+  ];
 
   return (
-    <View style={styles.container}>
-      <MapView
-        style={styles.map}
-        initialRegion={currentLocation || defaultRegion}
-        customMapStyle={darkModeStyle}
-        provider={PROVIDER_GOOGLE}
-        userInterfaceStyle={'dark'}
-      >
-        {markers.map((marker) => (
-          <Marker
-            key={marker.id}
-            coordinate={{
-              latitude: marker.latitude,
-              longitude: marker.longitude,
-            }}
-            title={marker.title}
-          />
-        ))}
-        {currentLocation && (
-          <Marker
-            coordinate={{
-              latitude: currentLocation.latitude,
-              longitude: currentLocation.longitude,
-            }}
-            title="You are here"
-            pinColor="blue"
-          />
-        )}
-      </MapView>
+    <View className='bg-black h-full'>
+       <MapViewComponent markers={markers} />
     </View>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  map: {
-    width: '100%',
-    height: '100%',
-  },
-});
