+++ harbor/assets/images/pointers/orgPointer.tsx
@@ -1,0 +1,34 @@
+import * as React from "react";
+
+const OrgPointer = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="67"
+    height="67"
+    fill="none"
+    viewBox="0 0 67 67"
+  >
+    <circle
+      cx="33.5"
+      cy="33.5"
+      r="33.5"
+      fill="url(#paint0_radial_6_3276)"
+      fillOpacity="0.5"
+    ></circle>
+    <defs>
+      <radialGradient
+        id="paint0_radial_6_3276"
+        cx="0"
+        cy="0"
+        r="1"
+        gradientTransform="rotate(90 0 33.5)scale(33.5)"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stopColor="#2F5DD3" stopOpacity="0"></stop>
+        <stop offset="1" stopColor="#2F5DD3"></stop>
+      </radialGradient>
+    </defs>
+  </svg>
+);
+
+export default OrgPointer;
