+++ harbor/types/index.ts
@@ -1,0 +1,8 @@
+export interface Marker {
+  id: number;
+  latitude: number;
+  longitude: number;
+  isSOS: boolean;
+  isVerified: boolean;
+  date: Date;
+}
